requirements.txt
puremagic
python-magic
androguard==3.4.0a1
rich
tqdm
colorama
oletools
pefile
pyaxmlparser
yara-python==4.5.0
prompt_toolkit==3.0.43
frida-tools
lief
pdfminer.six
requests
rarfile
acefile
Pillow
dpkt
pythonnet
pydnsbl
flare-floss
vivisect
pycryptodome
bs4
wh1tem0cha
psutil
pymem
pyja3